Split some parts of XmlUtils into XmlEncodingUtils so that parts can be moved into iaf-commons.
