+++ core/src/main/java/nl/nn/adapterframework/pipes/ReplacerPipe.java
@@ -26,7 +26,7 @@ import nl.nn.adapterframework.core.PipeRunResult;
 import nl.nn.adapterframework.doc.ElementType;
 import nl.nn.adapterframework.doc.ElementType.ElementTypes;
 import nl.nn.adapterframework.stream.Message;
-import nl.nn.adapterframework.util.XmlUtils;
+import nl.nn.adapterframework.util.XmlEncodingUtils;
 
 /**
  * Replaces all occurrences of one string with another.
@@ -108,9 +108,9 @@ public class ReplacerPipe extends FixedForwardPipe {
 		}
 		if (isReplaceNonXmlChars()) {
 			if (StringUtils.isEmpty(getReplaceNonXmlChar())) {
-				input = XmlUtils.stripNonValidXmlCharacters(input, isAllowUnicodeSupplementaryCharacters());
+				input = XmlEncodingUtils.stripNonValidXmlCharacters(input, isAllowUnicodeSupplementaryCharacters());
 			} else {
-				input = XmlUtils.replaceNonValidXmlCharacters(input, getReplaceNonXmlChar().charAt(0), false, isAllowUnicodeSupplementaryCharacters());
+				input = XmlEncodingUtils.replaceNonValidXmlCharacters(input, getReplaceNonXmlChar().charAt(0), false, isAllowUnicodeSupplementaryCharacters());
 			}
 		}
 		return new PipeRunResult(getSuccessForward(),input);
@@ -148,7 +148,7 @@ public class ReplacerPipe extends FixedForwardPipe {
 		lineSeparatorSymbol = string;
 	}
 
-	/** Replace all non XML chars (not in the <a href="http://www.w3.org/TR/2006/REC-xml-20060816/#NT-Char">character range as specified by the XML specification</a>) with {@link XmlUtils#replaceNonValidXmlCharacters(String, char, boolean, boolean) replaceNonValidXmlCharacters}
+	/** Replace all non XML chars (not in the <a href="http://www.w3.org/TR/2006/REC-xml-20060816/#NT-Char">character range as specified by the XML specification</a>) with {@link XmlEncodingUtils#replaceNonValidXmlCharacters(String, char, boolean, boolean) replaceNonValidXmlCharacters}
 	 * @ff.default false
 	 */
 	public void setReplaceNonXmlChars(boolean b) {
@@ -172,7 +172,7 @@ public class ReplacerPipe extends FixedForwardPipe {
 	}
 
 	/**
-	 * Whether to allow Unicode supplementary characters (like a smiley) during {@link XmlUtils#replaceNonValidXmlCharacters(String, char, boolean, boolean) replaceNonValidXmlCharacters}
+	 * Whether to allow Unicode supplementary characters (like a smiley) during {@link XmlEncodingUtils#replaceNonValidXmlCharacters(String, char, boolean, boolean) replaceNonValidXmlCharacters}
 	 * @ff.default false
 	 */
 	public void setAllowUnicodeSupplementaryCharacters(boolean b) {
