+++ core/src/main/java/nl/nn/adapterframework/util/XmlEncodingUtils.java
@@ -1,0 +1,250 @@
+package nl.nn.adapterframework.util;
+
+import org.apache.logging.log4j.Logger;
+
+public class XmlEncodingUtils {
+	static Logger log = LogUtil.getLogger(XmlEncodingUtils.class);
+	public static final char REPLACE_NON_XML_CHAR = 0x00BF; // Inverted question mark.
+	public static String encodeChars(String string) {
+		return encodeChars(string, false);
+	}
+
+	/**
+	 * Translates special characters to xml equivalents
+	 * like <b>&gt;</b> and <b>&amp;</b>. Please note that non valid xml chars
+	 * are not changed, hence you might want to use
+	 * replaceNonValidXmlCharacters() or stripNonValidXmlCharacters() too.
+	 */
+	public static String encodeChars(String string, boolean escapeNewLines) {
+		if (string==null) {
+			return null;
+		}
+		int length = string.length();
+		char[] characters = new char[length];
+		string.getChars(0, length, characters, 0);
+		return encodeChars(characters, 0, length, escapeNewLines);
+	}
+
+	public static String encodeCharsAndReplaceNonValidXmlCharacters(String string) {
+		return encodeChars(replaceNonValidXmlCharacters(string));
+	}
+
+	public static String encodeChars(char[] chars, int offset, int length) {
+		return encodeChars(chars, offset, length, false);
+	}
+
+	/**
+	 * Translates special characters to xml equivalents
+	 * like <b>&gt;</b> and <b>&amp;</b>. Please note that non valid xml chars
+	 * are not changed, hence you might want to use
+	 * replaceNonValidXmlCharacters() or stripNonValidXmlCharacters() too.
+	 */
+	public static String encodeChars(char[] chars, int offset, int length, boolean escapeNewLines) {
+
+		if (length<=0) {
+			return "";
+		}
+		StringBuilder encoded = new StringBuilder(length);
+		String escape;
+		for (int i = 0; i < length; i++) {
+			char c=chars[offset+i];
+			escape = escapeChar(c, escapeNewLines);
+			if (escape == null)
+				encoded.append(c);
+			else
+				encoded.append(escape);
+		}
+		return encoded.toString();
+	}
+
+	/**
+	 * Translates the five reserved XML characters (&lt; &gt; &amp; &quot; &apos;) to their normal selves
+	 */
+	public static String decodeChars(String string) {
+		StringBuilder decoded = new StringBuilder();
+
+		boolean inEscape = false;
+		int escapeStartPos = 0;
+
+		for (int i = 0; i < string.length(); i++) {
+			char cur=string.charAt(i);
+			if (inEscape) {
+				if ( cur == ';') {
+					inEscape = false;
+					String escapedString = string.substring(escapeStartPos, i + 1);
+					char unEscape = unEscapeString(escapedString);
+					if (unEscape == 0x0) {
+						decoded.append(escapedString);
+					}
+					else {
+						decoded.append(unEscape);
+					}
+				}
+			} else {
+				if (cur == '&') {
+					inEscape = true;
+					escapeStartPos = i;
+				} else {
+					decoded.append(cur);
+				}
+			}
+		}
+		if (inEscape) {
+			decoded.append(string.substring(escapeStartPos));
+		}
+		return decoded.toString();
+	}
+
+	/**
+	   * Conversion of special xml signs. Please note that non valid xml chars
+	   * are not changed, hence you might want to use
+	   * replaceNonValidXmlCharacters() or stripNonValidXmlCharacters() too.
+	   **/
+	private static String escapeChar(char c, boolean escapeNewLines) {
+		switch (c) {
+			case ('<') :
+				return "&lt;";
+			case ('>') :
+				return "&gt;";
+			case ('&') :
+				return "&amp;";
+			case ('\"') :
+				return "&quot;";
+			case ('\'') :
+				// return "&apos;"; // apos does not work in Internet Explorer
+				return "&#39;";
+			case ('\n')  :
+				if(escapeNewLines)
+					return "&#10;";
+		}
+		return null;
+	}
+
+	private static char unEscapeString(String str) {
+		if (str.equalsIgnoreCase("&lt;"))
+			return '<';
+		else if (str.equalsIgnoreCase("&gt;"))
+			return '>';
+		else if (str.equalsIgnoreCase("&amp;"))
+			return '&';
+		else if (str.equalsIgnoreCase("&quot;"))
+			return '\"';
+		else if (str.equalsIgnoreCase("&apos;") || str.equalsIgnoreCase("&#39;"))
+			return '\'';
+		else
+			return 0x0;
+	}
+
+	/**
+	 * Replaces non-unicode-characters by '0x00BF' (inverted question mark).
+	 */
+	public static int replaceNonPrintableCharacters(char[] buf, int offset, int len) {
+		if (len<0) {
+			return len;
+		}
+		int c;
+		int charCount = 0;
+		int counter = 0;
+		int readPos = 0;
+		int writePos = 0;
+		while(readPos<len) { // while no shift needs to be made, loop and replace where necessary
+			c=Character.codePointAt(buf, readPos+offset);
+			charCount = Character.charCount(c);
+			if (isPrintableUnicodeChar(c, true)) {
+				readPos += charCount;
+			} else {
+				buf[offset+readPos]= REPLACE_NON_XML_CHAR;
+				if (charCount==1) {
+					readPos++;
+				} else {
+					writePos = readPos+1;
+					readPos += charCount;
+					break;
+				}
+			}
+		}
+		while(readPos<len) { // continue with loop and shift after replacement was shorter than original
+			c=Character.codePointAt(buf, readPos+offset);
+			charCount = Character.charCount(c);
+			if (isPrintableUnicodeChar(c, true)) {
+				for(int j=0;j<charCount;j++) {
+					buf[offset+writePos++]=buf[offset+readPos++];
+				}
+			} else {
+				buf[offset+writePos++]= REPLACE_NON_XML_CHAR;
+				readPos+=charCount;
+				counter++;
+			}
+		}
+		if (counter>0 && log.isDebugEnabled()) log.debug("replaced ["+counter+"] non valid xml characters to ["+ REPLACE_NON_XML_CHAR+"] in char array of length ["+len+"]");
+		return writePos>0 ? writePos : readPos;
+	}
+
+	/**
+	 * Replaces non-unicode-characters by '0x00BF' (inverted question mark)
+	 * appended with #, the character number and ;.
+	 */
+	public static String replaceNonValidXmlCharacters(String string) {
+		return replaceNonValidXmlCharacters(string, REPLACE_NON_XML_CHAR, true, true);
+	}
+
+	public static String replaceNonValidXmlCharacters(String string, char to, boolean appendCharNum, boolean allowUnicodeSupplementaryCharacters) {
+		if (string==null) {
+			return null;
+		}
+		int length = string.length();
+		StringBuilder encoded = new StringBuilder(length);
+		int c;
+		int counter = 0;
+		for (int i = 0; i < length; i += Character.charCount(c)) {
+			c=string.codePointAt(i);
+			if (isPrintableUnicodeChar(c, allowUnicodeSupplementaryCharacters)) {
+				encoded.appendCodePoint(c);
+			} else {
+				if (appendCharNum) {
+					encoded.append(to + "#" + c + ";");
+				} else {
+					encoded.append(to);
+				}
+				counter++;
+			}
+		}
+		if (counter>0) {
+			if (log.isDebugEnabled()) log.debug("replaced ["+counter+"] non valid xml characters to ["+to+"] in string of length ["+length+"]");
+		}
+		return encoded.toString();
+	}
+
+	public static String stripNonValidXmlCharacters(String string, boolean allowUnicodeSupplementaryCharacters) {
+		int length = string.length();
+		StringBuilder encoded = new StringBuilder(length);
+		int c;
+		int counter = 0;
+		for (int i = 0; i < length; i += Character.charCount(c)) {
+			c=string.codePointAt(i);
+			if (isPrintableUnicodeChar(c,
+					allowUnicodeSupplementaryCharacters)) {
+				encoded.appendCodePoint(c);
+			} else {
+				counter++;
+			}
+		}
+		if (counter>0) {
+			if (log.isDebugEnabled()) log.debug("stripped ["+counter+"] non valid xml characters in string of length ["+length+"]");
+		}
+		return encoded.toString();
+	}
+
+	public static boolean isPrintableUnicodeChar(int c) {
+		return isPrintableUnicodeChar(c, false);
+	}
+
+	public static boolean isPrintableUnicodeChar(int c, boolean allowUnicodeSupplementaryCharacters) {
+		return (c == 0x0009)
+			|| (c == 0x000A)
+			|| (c == 0x000D)
+			|| (c >= 0x0020 && c <= 0xD7FF)
+			|| (c >= 0xE000 && c <= 0xFFFD)
+			|| (allowUnicodeSupplementaryCharacters && (c >= 0x00010000 && c <= 0x0010FFFF));
+	}
+}
